Move module and update tests
