+++ qiskit_ibm_runtime/fake_provider/executor/run_quantum_program.py
@@ -25,16 +25,16 @@ from qiskit.utils.optionals import HAS_AER
 from samplomatic import Tag
 from samplomatic.utils import get_annotation
 
-from ..quantum_program import CircuitItem, SamplexItem
-from ..results import QuantumProgramItemResult, QuantumProgramResult
+from ...quantum_program import CircuitItem, SamplexItem
+from ...results import QuantumProgramItemResult, QuantumProgramResult
 from .broadcast_sample import broadcast_sample
 from .insert_noise_pass import InsertNoisePass
 
 if TYPE_CHECKING:
     from qiskit.providers import BackendV2
 
-    from ..options_models.simulator import ExperimentalSimulatorOptions
-    from ..quantum_program import QuantumProgram
+    from ...options_models.simulator import ExperimentalSimulatorOptions
+    from ...quantum_program import QuantumProgram
 
 if HAS_AER:
     from qiskit_aer import AerSimulator
