+++ test/benchmarks/test_benchmark_executor_estimator.py
@@ -35,8 +35,8 @@ from qiskit_ibm_runtime.decoders.quantum_program.decoder import QuantumProgramRe
 from qiskit_ibm_runtime.executor_estimator.finalize_options import finalize_estimator_options
 from qiskit_ibm_runtime.executor_estimator.prepare import prepare
 from qiskit_ibm_runtime.executor_estimator.utils import find_unique_layers
-from qiskit_ibm_runtime.executor_local_mode.broadcast_sample import broadcast_sample
 from qiskit_ibm_runtime.fake_provider import FakeMarrakesh
+from qiskit_ibm_runtime.fake_provider.executor.broadcast_sample import broadcast_sample
 from qiskit_ibm_runtime.options_models.estimator import EstimatorOptions
 from qiskit_ibm_runtime.quantum_program import SamplexItem
 from qiskit_ibm_runtime.results.quantum_program import (
