+++ test/unit/local_mode/executor/test_broadcast_sample.py
@@ -18,11 +18,11 @@ from qiskit.transpiler import generate_preset_pass_manager
 from samplomatic.builders.build import build
 from samplomatic.transpiler import generate_boxing_pass_manager
 
-from qiskit_ibm_runtime.executor_local_mode.broadcast_sample import broadcast_sample
 from qiskit_ibm_runtime.fake_provider.backends.fez import FakeFez
+from qiskit_ibm_runtime.fake_provider.executor.broadcast_sample import broadcast_sample
 from qiskit_ibm_runtime.quantum_program import QuantumProgram
 
-from ...ibm_test_case import IBMTestCase
+from ....ibm_test_case import IBMTestCase
 
 
 class TestBroadcastSample(IBMTestCase):
