+++ test/unit/local_mode/executor/test_insert_noise_pass.py
@@ -22,9 +22,9 @@ from qiskit.quantum_info import DensityMatrix, PauliLindbladMap
 from qiskit.transpiler import PassManager
 from qiskit.utils import optionals
 
-from qiskit_ibm_runtime.executor_local_mode.insert_noise_pass import InsertNoisePass
+from qiskit_ibm_runtime.fake_provider.executor.insert_noise_pass import InsertNoisePass
 
-from ...ibm_test_case import IBMTestCase
+from ....ibm_test_case import IBMTestCase
 
 if optionals.HAS_AER:
     from qiskit_aer import AerSimulator
